Removed errors.rs in common
