+++ crates/client/Cargo.toml
@@ -11,6 +11,7 @@ repository.workspace = true
 prism-common.workspace = true
 prism-keys.workspace = true
 prism-serde.workspace = true
+prism-errors.workspace = true
 
 # webclient
 reqwest.workspace = true
