+++ verification_keys/keys.json
@@ -1,1 +1,1 @@
-{"base_vk": "0x00590b22eb65ab28d63742754a82fa1c9c868749c129e4edb7fa6e6e2022d099", "recursive_vk": "0x00d4431e19dd1a83d2fdbcb44ce8aa7de4953cbe6ebd9d5bf5e6c2e83a745765"}
+{"base_vk": "0x00367a2d1b4787e05884ba8d1ef6950c88345982c1fe4ce1a7863ae03b4a134f", "recursive_vk": "0x002dbd409e4f04acfe49dd099cbf4754f83ee6f1bd83e4305d2217318d644314"}
